fix: switch server assistant and session session unchanged
